Move `MainQueueDispatchDecorator` to seperate file
